backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
bcrypt
razorpay
python-dotenv
pydantic
httpx
google-genai